requirements.txt
astroid==1.6.5
certifi==2018.11.29
chardet==3.0.4
entrypoints==0.3
flake8==3.7.5
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.5.0
pycryptodome==3.7.3
pyflakes==2.1.0
pylint==1.9.4
requests==2.21.0
six==1.12.0
urllib3==1.24.1
wrapt==1.11.1
yapf==0.25.0
Django==2.2
django-bootstrap4==0.0.8
entrypoints==0.3
mccabe==0.6.1
pytz==2019.1
sqlparse==0.3.0
